# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: QuickServe - Beverage stand order management API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /events/{eventId}/orders:
    get:
      operationId: listOrders
      tags: [orders]
      summary: List orders for an event
      parameters:
        - name: eventId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: List of orders
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Order"
    post:
      operationId: createOrder
      tags: [orders]
      summary: Create a new order (buyer)
      parameters:
        - name: eventId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OrderInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
        "400":
          description: Validation error or duplicate name
        "409":
          description: Stock conflict
  /orders/{id}:
    get:
      operationId: getOrder
      tags: [orders]
      summary: Get order by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Order
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
  /events/{eventId}/orders/by-name/{name}:
    get:
      operationId: getOrderByName
      tags: [orders]
      summary: Get order by name (case-insensitive)
      parameters:
        - name: eventId
          in: path
          required: true
          schema:
            type: integer
        - name: name
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Order
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
        "404":
          description: Not found
  /orders/{id}/reserve:
    post:
      operationId: reserveOrder
      tags: [orders]
      summary: Reserve stock for an order
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Reserved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
        "409":
          description: Stock conflict
  /orders/{id}/pay:
    post:
      operationId: payOrder
      tags: [orders]
      summary: Process payment for a reserved order
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PaymentInput"
      responses:
        "200":
          description: Paid
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
        "400":
          description: Payment amount mismatch
        "409":
          description: Stock conflict
...
  /events/{eventId}/settings:
    get:
      operationId: getSettings
      tags: [settings]
      summary: Get event settings
      parameters:
        - name: eventId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Settings
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Settings"
    patch:
      operationId: updateSettings
      tags: [settings]
      summary: Update event settings
      parameters:
        - name: eventId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SettingsUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Settings"
...
components:
  schemas:
...
    OrderItem:
      type: object
      required: [id, commande_id, article_id, quantite, prix_unitaire, statut_livraison]
      properties:
        id:
          type: integer
        commande_id:
          type: integer
        article_id:
          type: integer
        article_nom:
          type: ["string", "null"]
        quantite:
          type: integer
        prix_unitaire:
          type: number
        statut_livraison:
          type: string
          enum: [non_livre, livre]
    Order:
      type: object
      required: [id, evenement_id, nom_commande, statut, montant_total, created_at, updated_at]
      properties:
        id:
          type: integer
        evenement_id:
          type: integer
        nom_commande:
          type: string
        statut:
          type: string
          enum: [en_attente, reservee, payee, livree_partiellement, livree, expiree]
        montant_total:
          type: number
        paye_cb:
          type: number
        paye_especes:
          type: number
        paye_cheque:
          type: number
        created_at:
          type: string
        updated_at:
          type: string
        expiration_reservation:
          type: ["string", "null"]
        items:
          type: array
          items:
            $ref: "#/components/schemas/OrderItem"
    OrderItemInput:
      type: object
      required: [article_id, quantite]
      properties:
        article_id:
          type: integer
        quantite:
          type: integer
    OrderInput:
      type: object
      required: [nom_commande, items]
      properties:
        nom_commande:
          type: string
        items:
          type: array
          items:
            $ref: "#/components/schemas/OrderItemInput"
    PaymentInput:
      type: object
      required: [paye_cb, paye_especes, paye_cheque]
      properties:
        paye_cb:
          type: number
        paye_especes:
          type: number
        paye_cheque:
          type: number
...
    Settings:
      type: object
      required: [id, evenement_id, temps_reservation_minutes, vente_ouverte, updated_at]
      properties:
        id:
          type: integer
        evenement_id:
          type: integer
        temps_reservation_minutes:
          type: integer
        vente_ouverte:
          type: boolean
        updated_at:
          type: string
    SettingsUpdate:
      type: object
      properties:
        temps_reservation_minutes:
          type: integer
        mdp_caisse:
          type: string
        mdp_preparateur:
          type: string
        mdp_admin:
          type: string
        vente_ouverte:
          type: boolean
...