# lib/api-zod/src/generated/types/order.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * QuickServe - Beverage stand order management API
 * OpenAPI spec version: 0.1.0
 */
import type { OrderItem } from './orderItem';
import type { OrderStatut } from './orderStatut';

export interface Order {
  id: number;
  evenement_id: number;
  nom_commande: string;
  statut: OrderStatut;
  montant_total: number;
  paye_cb?: number;
  paye_especes?: number;
  paye_cheque?: number;
  created_at: string;
  updated_at: string;
  /** @nullable */
  expiration_reservation?: string | null;
  items?: OrderItem[];
}