# lib/api-zod/src/generated/types/orderItem.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * QuickServe - Beverage stand order management API
 * OpenAPI spec version: 0.1.0
 */
import type { OrderItemStatutLivraison } from './orderItemStatutLivraison';

export interface OrderItem {
  id: number;
  commande_id: number;
  article_id: number;
  /** @nullable */
  article_nom?: string | null;
  quantite: number;
  prix_unitaire: number;
  statut_livraison: OrderItemStatutLivraison;
}