# lib/api-zod/src/generated/types/settings.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * QuickServe - Beverage stand order management API
 * OpenAPI spec version: 0.1.0
 */

export interface Settings {
  id: number;
  evenement_id: number;
  temps_reservation_minutes: number;
  vente_ouverte: boolean;
  updated_at: string;
}