# lib/api-zod/src/generated/types/settingsUpdate.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * QuickServe - Beverage stand order management API
 * OpenAPI spec version: 0.1.0
 */

export interface SettingsUpdate {
  temps_reservation_minutes?: number;
  mdp_caisse?: string;
  mdp_preparateur?: string;
  mdp_admin?: string;
  vente_ouverte?: boolean;
}